check existed convert

+++ luckyexcel4j-core/src/main/java/net/hanbd/luckyexcel4j/lucky/poi/base/CellValue.java
@@ -10,6 +10,7 @@ import net.hanbd.luckyexcel4j.lucky.poi.enums.CellHorizontalType;
 import net.hanbd.luckyexcel4j.lucky.poi.enums.CellVerticalType;
 import net.hanbd.luckyexcel4j.lucky.poi.enums.FontFamily;
 import net.hanbd.luckyexcel4j.lucky.poi.enums.TextBreakType;
+import org.apache.poi.xssf.usermodel.XSSFCellStyle;
 
 /**
  * 单元格数据
@@ -60,7 +61,7 @@ public class CellValue {
      * 字体大小
      */
     @JsonProperty("fs")
-    private Integer fontsize;
+    private Short fontsize;
     /**
      * 删除线.
      */
@@ -83,7 +84,8 @@ public class CellValue {
     @JsonProperty("mc")
     private MergeCell mergeCell;
     /**
-     * 竖排文字
+     * 竖排文字,<b>参考微软excel中 开始->对齐方式->方向</b>
+     * <p>
      * TODO 使用枚举
      *
      * <p>Text rotation,0: 0、1: 45 、2: -45、3 Vertical text、4: 90 、5: -90
@@ -91,17 +93,23 @@ public class CellValue {
     @JsonProperty("tr")
     private Integer textRotate;
     /**
-     * 文字旋转角度,取值范围[0-180]
+     * 文字旋转角度,取值范围[0-180].
+     * <p>
+     * 该字段与{@link CellValue#textRotate}其实表达的是一个意思,
+     * {@link CellValue#textRotate}是{@link CellValue#rotateText}表意的子集。
+     * 具体可查看{@link XSSFCellStyle#getRotation()}注释。
+     *
+     * @see XSSFCellStyle#getRotation()
      */
     @JsonProperty("rt")
-    private Integer rotateText;
+    private Short rotateText;
     /**
      * 文本换行. 0 截断、1溢出、2 自动换行
      */
     @JsonProperty("tb")
     private TextBreakType textBreak;
     /**
-     * 原始值.可能为字符串类型,也可能为数字类型. TODO 优化处理
+     * 原始值.可能为字符串类型,也可能为数字类型. TODO 值优化处理
      */
     @JsonProperty("v")
     private String value;
